Split investment transformers into seperate files
The transformers file was getting big and hard to find things, this makes it easier to maintain and avoid clashes betrween developer commits.
